electricityproject-main/python-api/requirements.txt
flask==2.2.3
mysql-connector-python==8.0.32
requests==2.28.2
pytz